Changed integer to Long

+++ src/main/java/org/example/JobSearch/dto/RespondedApplicantDTO.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @Builder
 public class RespondedApplicantDTO {
-    private Integer id;
-    private Integer resumeId;
-    private Integer vacancyId;
+    private Long id;
+    private Long resumeId;
+    private Long vacancyId;
     private Boolean confirmation;
 }
